Fixed index out of bound bug with player

+++ game/CaptainMath.py
@@ -181,15 +181,19 @@ class Player(spyral.Sprite):
             global ProwNum
 	    global PcolNum
 	    if (BoardStatus[ProwNum][PcolNum] != -1):
-	        if (ProwNum != 5):
+	        if (ProwNum != 4):
                     ProwNum+=1
             self.y = BoardYcoord[ProwNum][PcolNum]
         if (PcolNum == 6):
             PcolNum = 0
-            ProwNum+=1
         if (ProwNum == 5):
             ProwNum = 0
-
+        if (PcolNum == -1):
+            PcolNum = 5
+        if (ProwNum == -1):
+            ProwNum == 5
+        print ProwNum
+        print PcolNum
 class MathText(spyral.Sprite):
     def __init__(self, scene, index, answers, problem_question):
         spyral.Sprite.__init__(self, scene)
@@ -1016,7 +1020,7 @@ class CaptainMath(spyral.Scene):
           self.player.image = spyral.image.Image(filename = "images/mainPlayerRedImages/playerEnergyRight.png", size = None)
         if(playerLives == 1):
           self.player2.image = spyral.image.Image(filename = "images/mainPlayerRedImages/playerEnergyRight.png", size = None)
-        if(playerLives == 3):
+        if(playerLives == 0):
           self.player3.image = spyral.image.Image(filename = "images/mainPlayerRedImages/playerEnergyRight.png", size = None)
         pygame.mixer.init()
         FF = pygame.mixer.Sound("sounds/ohYeah.wav")
@@ -1148,7 +1152,7 @@ class CaptainMath(spyral.Scene):
                 self.arrow = Arrow(self)
                 print "gamestate = Levelselect"
 
-        if(rowNum == ProwNum and colNum == PcolNum and isplayerDead == False):
+        if(rowNum == ProwNum and colNum == PcolNum and isplayerDead == False and forceFieldOn == False):
             if(playerLives == 2):
                 self.player.kill()
             elif(playerLives == 1):
